Use clamp(x,0,1) instead of min(x,1)/max(x,0) where possible
So, this is a microoptimization, but I might as well do it. Probably didn't gain even one frame from this.

GPUs for a very long time now have had saturate(x) from HLSL implemented in hardware which is equivalent to clamp(x, 0, 1) and can be optimized into by the driver.

min() and max() are usually done with branching, which is slower.