Did some work on lighting
I've moved water caustics rendering to be able to do it for stained glass as well, and i've done a few optimizations to reduce the performance hit this has. I've also gotten something to be able to have water fog absorb sunlight properly, disabled by default tough as it doesn't work very well.

I'm considering moving RSM to be able to do it on stained glass as well, it would lose the very importany filtering tough but I guess if I finish up TAA and make TAA be on by default the temporal filtering effect it has would be enough.