Added kinda WIP terrain parallax and did some fixes and tweaks
This is a different method for terrain parallax to what other shaderpacks use.

It has the advantage of being exact, but the disadvantage of potentially being very slow (especially with high resolution textures).

I intend to add a more standard version that's not quite as accurate, but way faster for high resolution textures.