Refactored shadows, significantly improved HBAO
HBAO still needs improvements, but once I'm done with the little bit that's left I'll be removing the other AO methods as HBAO is fairly close in realism to RTAO while requiring a lot fewer samples than SSAO to get rid of most of the noise, even with a significantly larger radius.

Issue right now is that it's less accurate the further away from the center it is, but aside from that it's very good.