Various changes
Skylight color is now based on the atmosphere directly up, this is more accurate and is darker than before, making RSM much more noticable. This is NOT implemented in an efficient way right now, so performance hit will be noticable.

HBAO has gotten a fix for some overocclusion that occurs mainly far away, which also has a performance hit.

I've optimized RSM a little tough, and autoexposure has more range in dark scenes so it's a little easier to see at night.