Made RSM filter depth-aware, store RSM in colortex2
Reason I moved it to be stored in colortex2 is to hopefully lower memory 
bandwidth a little, but also to allow me to possibly do a filter over 
multiple passes. I'm not sure I'll actually do that, but I know that 
doing that can give very high quality while still being decently fast so 
I want to at least try it.