Brighter night sky, darker blocklight, tweaked chroma gamma
Brightening the night sky like this is unrealistic, but it lets me decrease the max exposure and make nighttime appear darker overall without making it too hard to see. There's an option to change it back to realistic levels, if you prefer that. Max exposure and ambient light is also adjusted with that option, in order to make sure it's not too hard to see.