Optimized cloud shadows
This works by rendering a small shadow map for just cloud shadows into 
one of the buffers.
Currently it isn't used for distant VL, which would benefit from this a 
lot, but I'd need another program for that and right now I don't want to 
do so just to optimize a single effect that probably isn't even going to 
be used that often outside of screenshots.