Smaller default waves, more useful options for waves
I find that these waves look a bit more appropriate in most cases.

For the options, instead of setting the wave height directly, it's now 
set as a ratio relative to the wavelength (divided by tau). This makes 
it a bit easier to get the intended look and scale.
There's also a new option to adjust the sharpness of the waves. This can 
be useful to make the waves seem taller than they actually are. The 
older waves used a value approximately equivalent te 0.65, but the new 
ones have it set to 0 - which is probably more realistic anyway.