Changed default autoexposure method back to "Simple".
The histogram-based autoexposure I was using doesn't really work as well 
as I'd like it to.
I really like how it currently looks at night - that's the main reason 
I've kept it as the default for so long - but during the day I find that 
things often end up a bit darker than I'd like.
If I get around to improving it at some point I might change the default 
back to it.