Big caustics improvement, at cost of RSM leak prevention
I now encode projected caustics for any depth in the shadow map.
This allows me to re-use this information for the "high" caustics setting, with no accuracy loss, and providing vastly more detail at the same quality setting!

Had to remove RSM leak prevention for this to fit, though, as I needed another channel in the shadow map to store this information.
Overall I think it's definitely worth it though.