Fast high-quality refractive caustics
The key algorithm is searching for the source of the caustic via Newton-Rhapson iteration, and then figures out how bright that caustic should be using the projected caustics algorithm. Yeah, it's that simple. Should have figured this one out years ago.

Of course, the limitation this algorithm is that it only finds a single source, so on its own it can't solve more complex cases like folds where there are multiple caustic sources.

The neat part is that my previous high-quality algorithm basically just gives an estimate of where each caustic source is and leaves brightness to another method, so it now uses this new algorithm to refine those estimates, allowing it to produce much better results.